isolate aggregation disruption junit calculation
